// app/src/app/Models/PersonaJuridicaViewModel.ts
export class PersonaJuridica {
  Id: number;
  RtnSolicitante: string;

  esCi_RepresentanteLegal: string;
  esCi_Id_RepresentanteLegal: number;
  Profesion_RepresentanteLegal: string;
  prof_Id_RepresentanteLegal: number;

  Estado: string;
  esta_Id: number;
  Ciudad: string;
  ciud_Id: number;
  DireccionCompleta: string;

  Estado_RepresentanteLegal: string;
  esta_Id_RepresentanteLegal: number;
  Ciudad_RepresentanteLegal: string;
  ciud_Id_RepresentanteLegal: number;
  DireccionCompleta_RepresentanteLegal: string;

  TelefonoFijoEmpresa: string;
  TelefonoFijo: string;
  TelefonoCelular: string;
  Correo: string;
  CorreoAlternativo: string;
  Aduana: string;
  adua_Id: number;

  Rtn: string;
  RtnUrl: string;
  Rtn_RepresentanteLegal: string;
  RtnUrl_RepresentanteLegal: string;
  Dni_RepresentanteLegal: string;
  DniUrl_RepresentanteLegal: string;
  EscrituraPublica: string;
  EscrituraPublicaUrl: string;
}